% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{boxplotb}
\alias{boxplotb}
\title{Bivariate boxplot}
\usage{
boxplotb(handle, Y, ...)
}
\arguments{
\item{handle}{Engine handle from \code{\link[=start_engine]{start_engine()}}.}

\item{Y}{Numeric matrix with two columns.}

\item{...}{Optional arguments for \code{boxplotb}.}
}
\value{
A list with centroid, contour and outlier indices.
}
\description{
Bivariate boxplot
}
\examples{
\dontrun{
h = start_engine("boxplotb")
out = boxplotb(h, Y)
stop_engine(h)
}
}
