% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tclust.R
\name{fsda_tclust}
\alias{fsda_tclust}
\title{Run FSDA tclust clustering}
\usage{
fsda_tclust(data, k, alpha, restrfactor = NULL, ...)
}
\arguments{
\item{data}{Matrix of observations}

\item{k}{Number of clusters}

\item{alpha}{Trimming level}

\item{restrfactor}{Restriction factor on scatter matrices}

\item{...}{Additional FSDA MATLAB name-value options}
}
\value{
An object of class \code{"fsda_tclust"}
}
\description{
Run FSDA tclust clustering
}
