% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LXS.R
\name{LXS}
\alias{LXS}
\title{LXS}
\usage{
LXS(handle, y, X, ...)
}
\arguments{
\item{handle}{MATLAB engine handle.}

\item{y}{Response variable vector.}

\item{X}{Predictor matrix.}

\item{...}{Additional arguments passed to MATLAB.}
}
\value{
MATLAB engine output for LXS.
}
\description{
Computes least trimmed squares (LXS) estimators.
}
