% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{malfwdplot}
\alias{malfwdplot}
\title{Mahalanobis distance trajectories from a forward search}
\usage{
malfwdplot(handle, out, ...)
}
\arguments{
\item{handle}{Engine handle from \code{\link[=start_engine]{start_engine()}}.}

\item{out}{Output of \code{FSMeda}.}

\item{...}{Optional arguments for \code{malfwdplot}.}
}
\value{
\code{NULL}.
}
\description{
Mahalanobis distance trajectories from a forward search
}
\examples{
\dontrun{
h = start_engine("malfwdplot")
malfwdplot(h, out)
stop_engine(h)
}
}
