% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{render_figures}
\alias{render_figures}
\title{Render pending MATLAB figures}
\usage{
render_figures(handle)
}
\arguments{
\item{handle}{Engine handle from \code{\link[=start_engine]{start_engine()}}.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Render pending MATLAB figures
}
