% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{yXplot}
\alias{yXplot}
\title{Plot y against each column of X}
\usage{
yXplot(handle, y, X, ...)
}
\arguments{
\item{handle}{Engine handle from \code{\link[=start_engine]{start_engine()}}.}

\item{y}{Response.}

\item{X}{Predictors.}

\item{...}{Optional arguments for \code{yXplot}.}
}
\value{
\code{NULL}.
}
\description{
Plot y against each column of X
}
\examples{
\dontrun{
h = start_engine("yXplot")
yXplot(h, y, X)
stop_engine(h)
}
}
